Implement 'mkNeuralNetwork' (ignoring Nguyen-Widrow)
Nguyen-Widrow initialization method is left TODO for now, basic and
normalized initialization methods are provided, which is enough for
a start.

Some additional type synonyms are provided for signature readability.

To allow fast prototyping, default values are provided, which should
allow anybody to quickly start using the library, these include:

 - defaultInitStrategy which defaults to normalized

 - defaultParameters which defaults to:

        * fixed learning rate of 1.0

        * mini-batch gradient descent, size 10

        * cross-entropy cost function